apk: split `apk` out of `alpine`
Refs: https://github.com/quay/claircore/issues/923
