+++ apk/scanner_test.go
@@ -1,4 +1,4 @@
-package alpine
+package apk
 
 import (
 	"context"
